Changed header guard for format.h
Previously, if spdlog was being used in a project that also uses cppformat, this file (with it's custom namespace) would not be included.  This is because, while the definitions were added to the spdlog namespace, the include guard was left the same as in the actual cppformat library.  This change modifies the include guard by prefixing it with SPDLOG_.