fix broken examples in run from arbitrary dirs
examples work fine when run from example/, as in

 $ cd example/
 $ ./example

but otherwise break:

 $ cd
 $ ~/path/to/spdlog/example/bench

the above would fail on trying to open the logs/somefile for writing if
the logs directory doesn't exist in the current directory.

let's not assume there is a logs/ directory and just write our logs in
the current directory. we can suffix the logs with .log so they're easy
to find in .gitignore for cleanup.

and with that, we don't need the logs direcory, or the .gitignore inside
it that keeps it hanging.