Symbol visiblity
The commits defines a SPDLOG_API to mark classes that should
be exported when built as part of a shared library.
Only the registry_t class is marked with SPDLOG_API for now.

A new entry in tweakme.h has been added,
SPDLOG_ENABLE_DLL_IMPORT_EXPORT which default to undefined.
If defined, the import / export mechanism will be enabled and
if SPDLOG_EXPORTS is defined, the SPDLOG_API will export the
symbols, otherwise it will import them.

The default behavior is unchanged and SPDLOG_API will
expand to nothing.

See #693