Changes according to review
We need first clang-tidy for all

If found, we should find run-clang-tidy.py for the check taget
Then we need a compile_commands.json created by cmake or mesonbuild.

And this has to be linkd to the source tree, if the binary tree is
outsite the source tree.

That must be true for ninja generator to generate absolute pathes at
compile_commands.json!