Add an option that allow user to disable warnings
This option is false by default, thus do not change the current behavior
of the build system.

This is useful when the user vendor spdlog's code into his own
codebase and do not want to have the compiler warn about spdlog's code.

This also is an easy way to prevent CMake to set non-compatible flags on
`clang-cl` (clang for Visual Studio) like `-pedantic`.

In practice, this patch does:

 -add option `SPDLOG_DISABLE_WARNINGS`, `off` state by default
 -add conditional test checking that the option above is `off`.

If the option is manually turned on by the user: behavior is to
skip the call to `spdlog_enable_warnings` on `spdlog` build target.

If the user does nothing, current behavior is unchanged.