Add %N pattern formatter for logging the thread name using pthreads
This commit adds the SPDLOG_NO_PTHREAD_ID CMake option for enabling
saving of the pthread handle. This can be used in conjunction with the
%N pattern formatter for logging the name of the thread. This can be
really useful if you have a large application with many threads that are
hard to differentiate.

The option is disabled by default as I'm sure this won't work on
windows.

This patch also adds a test for testing the new flag.