cmake: prevent leak of libfmt dependency into spdlog users
When linking dynamically against spdlog there is no need for the user of
the shared library to itself link against the dependencies of that
spdlog. Doing it anyway causes problems when one of the dependencies
gets a soname bump, because now the library user needs to be rebuild
despite not using a single symbol of the updated dependency.